Refactor PublicProfile component to fetch and display user profile, posts, and public reading lists, enhancing user interaction and data presentation
